+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Upcoming Changes (unreleased)
+- added ```valid_country_name``` method to return country name where the parsed phone comes from (in case it's valid)
 
 ## 0.6.26 - October 15, 2018
 - updated data